+++ src/sensors/AtlasScientificDO.h
@@ -288,8 +288,7 @@ class AtlasScientificDO_DOmgL : public Variable {
     explicit AtlasScientificDO_DOmgL(
         AtlasScientificDO* parentSense, const char* uuid = "",
         const char* varCode = ATLAS_DOMGL_DEFAULT_CODE)
-        : Variable(parentSense, static_cast<uint8_t>(ATLAS_DOMGL_VAR_NUM),
-                   static_cast<uint8_t>(ATLAS_DOMGL_RESOLUTION),
+        : Variable(parentSense, ATLAS_DOMGL_VAR_NUM, ATLAS_DOMGL_RESOLUTION,
                    ATLAS_DOMGL_VAR_NAME, ATLAS_DOMGL_UNIT_NAME, varCode, uuid) {
     }
     /**
@@ -322,8 +321,7 @@ class AtlasScientificDO_DOpct : public Variable {
     explicit AtlasScientificDO_DOpct(
         AtlasScientificDO* parentSense, const char* uuid = "",
         const char* varCode = ATLAS_DOPCT_DEFAULT_CODE)
-        : Variable(parentSense, static_cast<uint8_t>(ATLAS_DOPCT_VAR_NUM),
-                   static_cast<uint8_t>(ATLAS_DOPCT_RESOLUTION),
+        : Variable(parentSense, ATLAS_DOPCT_VAR_NUM, ATLAS_DOPCT_RESOLUTION,
                    ATLAS_DOPCT_VAR_NAME, ATLAS_DOPCT_UNIT_NAME, varCode, uuid) {
     }
     /**
