+++ src/sensors/AtlasScientificpH.h
@@ -238,9 +238,8 @@ class AtlasScientificpH_pH : public Variable {
     explicit AtlasScientificpH_pH(AtlasScientificpH* parentSense,
                                   const char*        uuid = "",
                                   const char* varCode = ATLAS_PH_DEFAULT_CODE)
-        : Variable(parentSense, static_cast<uint8_t>(ATLAS_PH_VAR_NUM),
-                   static_cast<uint8_t>(ATLAS_PH_RESOLUTION), ATLAS_PH_VAR_NAME,
-                   ATLAS_PH_UNIT_NAME, varCode, uuid) {}
+        : Variable(parentSense, ATLAS_PH_VAR_NUM, ATLAS_PH_RESOLUTION,
+                   ATLAS_PH_VAR_NAME, ATLAS_PH_UNIT_NAME, varCode, uuid) {}
     /**
      * @brief Destroy the AtlasScientificpH_pH object - no action needed.
      */
